+++ changelog.md
@@ -1,6 +1,11 @@
 Project-Stats - Changelog
 =========================================
 
+- **Project-Stats 0.4** - (09/24/2013)
+    - Added more compatibility with Python 3.
+	- Added new file types.
+	- Minor changes.
+
 - **Project-Stats 0.3** - (08/26/2013)
 	- Added Python 3 Compatibility.
 	- GUI Simplification.